refactor(main-code): create Jump.tsx
