+++ manifest.json
@@ -1,7 +1,7 @@
 {
   "name": "Jump",
   "id": "1009330872280993883",
-  "api": "1.1.0",
+  "api": "1.0.0",
   "main": "dist/code.js",
   "ui": "dist/ui.html"
 }
